add provider methods, move function from index
https://coveord.atlassian.net/browse/KIT-2606
